Relocate main to root directory
