+++ .goreleaser.yml
@@ -4,7 +4,7 @@ before:
 builds:
 - id: npmi-go
   binary: npmi-go
-  main: ./cmd/npmi-go/main.go
+  main: main.go
   env:
     - CGO_ENABLED=0
   goos:
