+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/hermo/npmi-go/cmd"
+
+func main() {
+	cmd.Execute()
+}
